moved directories around, just made it cleaner to access. And made vision directory.
