--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OldFiles/FinalAuto.java
@@ -1,161 +1,0 @@
-package org.firstinspires.ftc.teamcode.OldFiles;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.openftc.easyopencv.OpenCvCamera;
-import org.openftc.easyopencv.OpenCvCameraFactory;
-import org.openftc.easyopencv.OpenCvCameraRotation;
-import org.openftc.easyopencv.OpenCvWebcam;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import static android.os.SystemClock.sleep;
-
-/*
-/*
- * Copyright (c) 2021 OpenFTC Team
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-
-
-/* * This sample demonstrates how to run analysis during INIT
- * and then snapshot that value for later use when the START
- * command is issued.*/
-@TeleOp(name = "Final Autonomous")
-public class FinalAuto extends LinearOpMode
-{
-
-//    DcMotor lf, lb, rf, rb;
-    OpenCvWebcam webcam;
-    SleeveDetection pipeline;
-    SleeveDetection.ParkingPosition snapshotAnalysis = SleeveDetection.ParkingPosition.LEFT;// default
-
-    @Override
-    public void runOpMode()
-    {
-
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        webcam = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
-        pipeline = new SleeveDetection();
-        webcam.setPipeline(pipeline);
-
-        webcam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
-        {
-            @Override
-            public void onOpened()
-            {
-                webcam.startStreaming(320,240, OpenCvCameraRotation.UPRIGHT); // 320x240 for lower res and higher performance
-            }
-
-            @Override
-            public void onError(int errorCode) {}
-        });
-
-        /*
-         * The INIT-loop:
-         * This REPLACES waitForStart!
-         */
-        while (!isStarted() && !isStopRequested())
-        {
-//            lf = hardwareMap.dcMotor.get("lf");
-//            lb = hardwareMap.dcMotor.get("lb");
-//            rf = hardwareMap.dcMotor.get("rf");
-//            rb = hardwareMap.dcMotor.get("rb");
-//
-//            DcMotor[] motors = {lf, lb, rf, rb};
-//
-//            for(DcMotor motor : motors) {
-//                motor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-//                motor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-//                motor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-//            }
-//            telemetry.addData("Realtime analysis", pipeline.getPosition());
-
-            // Don't burn CPU cycles busy-looping in this sample
-            sleep(50);
-        }
-
-        /*
-         * The START command just came in: snapshot the current analysis now
-         * for later use. We must do this because the analysis will continue
-         * to change as the camera view changes once the robot starts moving!
-         */
-        snapshotAnalysis = pipeline.getPosition();
-        /*
-         * Show that snapshot on the telemetry
-         */
-        telemetry.addData("Snapshot post-START analysis", snapshotAnalysis);
-        telemetry.update();
-
-        waitForStart();
-
-        switch (snapshotAnalysis)
-        {
-            case LEFT:
-            {
-//                rf.setPower(.5);
-//                rb.setPower(-.5);
-//                lf.setPower(-.5);
-//                lb.setPower(.5);
-                telemetry.addData("Park Left", "LEFT");
-                telemetry.update();
-                sleep(3000);
-                break;
-            }
-
-            case RIGHT:
-            {
-//                rf.setPower(-.5);
-//                rb.setPower(.5);
-//                lf.setPower(.5);
-//                lb.setPower(-.5);
-//                sleep(3000);
-                telemetry.addData("Park Right", "RIGHT");
-                telemetry.update();
-                sleep(3000);
-                break;
-            }
-
-            case CENTER:
-            {
-                /* Your autonomous code*/
-                telemetry.addData("Park Center", "CENTER");
-                telemetry.update();
-                sleep(3000);
-
-                break;
-            }
-        }
-//        lf.setPower(.5);
-//        lb.setPower(.5);
-//        rf.setPower(.5);
-//        rb.setPower(.5);
-
-        sleep(2000);
-
-        /* You wouldn't have this in your autonomous, this is just to prevent the sample from ending */
-        while (opModeIsActive())
-        {
-            // Don't burn CPU cycles busy-looping in this sample
-            sleep(50);
-        }
-    }
-}
-
