+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/SimpleArmCommand.java
@@ -1,7 +1,7 @@
-package org.firstinspires.ftc.teamcode.FTCLib_Programs.command;
+package org.firstinspires.ftc.teamcode.command;
 
 import com.arcrobotics.ftclib.command.CommandBase;
-import org.firstinspires.ftc.teamcode.FTCLib_Programs.subsystem.ArmSubsystem;
+import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 public class SimpleArmCommand extends CommandBase {
     protected final ArmSubsystem arm;
