+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/lift/MoveLiftTo.java
@@ -1,7 +1,7 @@
-package org.firstinspires.ftc.teamcode.FTCLib_Programs.command.lift;
+package org.firstinspires.ftc.teamcode.command.lift;
 
-import org.firstinspires.ftc.teamcode.FTCLib_Programs.command.SimpleArmCommand;
-import org.firstinspires.ftc.teamcode.FTCLib_Programs.subsystem.ArmSubsystem;
+import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
+import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 public class MoveLiftTo extends SimpleArmCommand {
     int targetPos;
