+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/GetHeight.java
@@ -1,14 +1,11 @@
-package org.firstinspires.ftc.teamcode.OldFiles;
+package org.firstinspires.ftc.teamcode.opmode;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.Gamepad;
-import com.qualcomm.robotcore.hardware.Servo;
 
-@TeleOp(name = "getHeight")
-public class getHeight extends OpMode{
+@TeleOp(name = "GetHeight")
+public class GetHeight extends OpMode{
 
     DcMotor lift;
