+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/WIP_TeleOp.java
@@ -1,13 +1,12 @@
-package org.firstinspires.ftc.teamcode.FTCLib_Programs.opmode;
+package org.firstinspires.ftc.teamcode.opmode;
 
 import com.arcrobotics.ftclib.command.button.Button;
 import com.arcrobotics.ftclib.command.button.GamepadButton;
 import com.arcrobotics.ftclib.gamepad.GamepadEx;
 import com.arcrobotics.ftclib.gamepad.GamepadKeys;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import org.firstinspires.ftc.teamcode.FTCLib_Programs.command.drive.DefaultFieldCentricDrive;
-import org.firstinspires.ftc.teamcode.FTCLib_Programs.command.drive.DefaultRobotCentricDrive;
+import org.firstinspires.ftc.teamcode.command.drive.DefaultFieldCentricDrive;
+import org.firstinspires.ftc.teamcode.command.drive.DefaultRobotCentricDrive;
 
 @TeleOp(name = "BasicDrive")
 public class WIP_TeleOp extends BaseOpMode {
