+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystem/ArmSubsystem.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.FTCLib_Programs.subsystem;
+package org.firstinspires.ftc.teamcode.subsystem;
 
 
 import com.arcrobotics.ftclib.command.SubsystemBase;
